requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==1.10.7
email-validator==2.0.0
cryptography==40.0.2
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.1.0